+++ Sources/SwifterSwift/Foundation/NSAttributedStringExtensions.swift
@@ -180,4 +180,27 @@ public extension NSAttributedString {
     }
 }
 
+public extension Array where Element: NSAttributedString {
+    /// SwifterSwift: Returns a new `NSAttributedString` by concatenating the elements of the sequence, adding the given separator between each element.
+    ///
+    /// - Parameter separator: An `NSAttributedString` to add between the elements of the sequence.
+    /// - Returns: NSAttributedString with applied attributes.
+    // https://stackoverflow.com/questions/32830519/is-there-joinwithseparator-for-attributed-strings
+    func joined(separator: NSAttributedString) -> NSAttributedString {
+        guard let firstElement = first else { return NSMutableAttributedString(string: "") }
+        return dropFirst().reduce(into: NSMutableAttributedString(attributedString: firstElement)) { result, element in
+            result.append(separator)
+            result.append(element)
+        }
+    }
+
+    func joined(separator: String) -> NSAttributedString {
+        guard let firstElement = first else { return NSMutableAttributedString(string: "") }
+        let attributedStringSeparator = NSAttributedString(string: separator)
+        return dropFirst().reduce(into: NSMutableAttributedString(attributedString: firstElement)) { result, element in
+            result.append(attributedStringSeparator)
+            result.append(element)
+        }
+    }
+}
 #endif